Update all controllers

+++ controller/Comments/CommentsController.js
@@ -5,16 +5,15 @@ const usersTable = db['Users'];
 const getComment = async (req, res) => {
     try {
         const comment = await commentsTable.findByPk(req.params.id);
-        if (comment !== null) {
+        if (comment) {
             res.status(200).send({
                 data: comment
             })
         } else {
-            res.status(404).send({
-                message: 'Comment was not found'
+            res.status(422).send({
+                message: 'Comment pas trouvé'
             })
         }
-
     } catch (error) {
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
@@ -22,15 +21,18 @@ const getComment = async (req, res) => {
         })
     }
 }
-
 const getComments = async (req, res) => {
     try {
         const comments = await commentsTable.findAll();
-
-        res.status(200).send({
-            data: comments
-        })
-
+        if (comments && comments.length !== 0) {
+            res.status(200).send({
+                data: comments
+            })
+        } else {
+            res.status(422).send({
+                message: 'Pas de commentaire trouvé'
+            })
+        }
     } catch (error) {
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
@@ -38,28 +40,26 @@ const getComments = async (req, res) => {
         })
     }
 }
-
 const createComment = async (req, res) => {
     try {
-        let data = { ...req.body };
-        let idUserFound = await usersTable.findOne({
-            where: { id: data.idUsers }
-        });
-        if (idUserFound !== null) {
+        const data = { ...req.body };
+        const idUser = data.idUsers;
+        const idUserFound = await usersTable.findByPk(idUser)
+        if (idUserFound) {
             const newComments = await commentsTable.create(data);
-            if (newComments !== null) {
+            if (newComments) {
                 res.status(200).send({
-                    message: 'Comment created',
+                    message: 'Commentaire créé',
                     data: newComments
                 })
             } else {
-                res.status(400).send({
-                    message: 'Comment was not created'
+                res.status(422).send({
+                    message: 'Commentaire pas créé'
                 })
             }
         } else {
-            res.status(404).send({
-                message: 'idUsers not existed'
+            res.status(422).send({
+                message: 'Utilisateur pas trouvé'
             })
         }
     } catch (error) {
@@ -69,26 +69,29 @@ const createComment = async (req, res) => {
         })
     }
 }
-
 const modifyComment = async (req, res) => {
     try {
         const { comment } = req.body;
-        const idComments = req.params.id;
-        const updateComments = await commentsTable.update(
-            { comment: comment },
-            {
-                where: {
-                    id: idComments
-                }
-            })
-        if (updateComments[0] == 1) {
-            res.status(200).send({
-                message: 'Comment was updated'
-            });
-        }
-        else {
-            res.status(400).send({
-                message: 'Comment was not updated'
+        const idComment = req.params.id;
+        const idCommentFound = await commentsTable.findByPk(idComment);
+        if (idCommentFound) {
+            const updateComments = await commentsTable.update(
+                { comment: comment },
+                {
+                    where: { id: idComment }
+                })
+            if (updateComments[0] == 1) {
+                res.status(200).send({
+                    message: 'Commentaire modifié'
+                });
+            } else {
+                res.status(422).send({
+                    message: 'Commentaire pas modifié'
+                })
+            }
+        } else {
+            res.status(422).send({
+                message: 'Commentaire pas trouvé'
             })
         }
     } catch (error) {
@@ -98,21 +101,26 @@ const modifyComment = async (req, res) => {
         })
     }
 }
-
 const deleteComment = async (req, res) => {
     try {
         const idComment = req.params.id;
-        const deletedComment = await commentsTable.destroy({
-            where: { id: idComment }
-        })
-        console.log(deletedComment);
-        if (deletedComment == 1) {
-            res.status(200).send({
-                message: 'Comment was deleted'
+        const idCommentFound = await commentsTable.findByPk(idComment);
+        if (idCommentFound) {
+            const deletedComment = await commentsTable.destroy({
+                where: { id: idComment }
             })
+            if (deletedComment == 1) {
+                res.status(200).send({
+                    message: 'Comment supprimé'
+                })
+            } else {
+                res.status(422).send({
+                    message: 'Comment pas supprimé'
+                })
+            }
         } else {
-            res.status(400).send({
-                message: 'Comment was not deleted'
+            res.status(422).send({
+                message: 'Commentaire pas trouvé'
             })
         }
     } catch (error) {
